+++ tests/fixtures/jupytext/readme_basic_percent.py
@@ -1,0 +1,9 @@
+# %% [markdown]
+# This is a markdown cell
+
+# %%
+def f(x):
+    return 3 * x + 1
+
+# %%
+f(3)